Fixed table speakers and talks.
Former-commit-id: 31a497b84132033690cefaa8cf17143cbeba2499
Former-commit-id: 08b5e3c20d2dca3b55919406535e0928ab7e45b5

+++ generate_pdf_talk.php
@@ -189,7 +189,9 @@ $outdir = __DIR__ . "/data";
 $texFile = $outdir . '/' . $outfile . ".tex";
 $pdfFile = $outdir . '/' . $outfile . ".pdf";
 
-unlink( $pdfFile );
+if( file_exists( $pdfFile ) )
+    unlink( $pdfFile );
+
 file_put_contents( $texFile,  $TeX );
 $cmd = "pdflatex --output-directory $outdir $texFile";
 if( file_exists( $texFile ) )
